Put Deploy code in a position to be tested
