+++ src/Modules/Dependencies.php
@@ -4,6 +4,8 @@ namespace Dxw\Whippet\Modules;
 
 class Dependencies extends \RubbishThorClone
 {
+    use Common;
+
     public function __construct()
     {
         parent::__construct();
@@ -17,21 +19,6 @@ class Dependencies extends \RubbishThorClone
         $this->command('install', 'Installs dependencies');
         $this->command('update', 'Updates dependencies to their latest versions');
         $this->command('migrate', 'Converts legacy plugins file to whippet.json');
-    }
-
-    private function exitIfError(\Result\Result $result)
-    {
-        if ($result->isErr()) {
-            echo sprintf("ERROR: %s\n", $result->getErr());
-            exit(1);
-        }
-    }
-
-    private function getDirectory()
-    {
-        $this->exitIfError($this->projectDirectory);
-
-        return $this->projectDirectory->unwrap();
     }
 
     public function install()
